Move site helpers to a separate file
